# openapi/auth-service-v1.yaml
openapi: 3.0.3
info:
  title: Craves Auth Service API
  version: 1.0.0
  description: Authentication endpoints for Firebase phone token exchange and Craves access/refresh session management.
servers:
  - url: https://api.craves.in
paths:
...
  /api/v1/auth/me:
    get:
      summary: Get current authenticated Craves identity
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Access token missing, invalid, or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MeResponse:
      type: object
      properties:
        identity:
          $ref: '#/components/schemas/IdentityResponse'
    IdentityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firebaseUid:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        displayName:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        roles:
          type: array
          items:
            type: string
            enum: [CUSTOMER, CHEF, ADMIN]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time